:sparkles: feat: Adição do segundo endpoint para busca de queimadas
